New primitives [block], [unblock] and [run].
This also introduces a new [Blocked T] type, and requires extending
the types allowed for primitives.

A new [fconstr] constructor for primitives is added. It represents a
primitive applied to all its suitably-evaluated arguments. Note that
it can either be:
- Ready to be evaluated, if the primitive has a reduction rule.
- A value, if it does not (like [block]).

These two cases are distinguished by setting the mark to [Cnst] when
it is discovered that a primitive is a value.